requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain==0.0.350
langchain-community==0.0.4
chromadb==0.4.18
anthropic==0.7.7
pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
structlog==23.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
